fetch resources in sequence avoid state wiped out
