+++ Composer/packages/client/src/components/CreationFlow/CreationFlow.tsx
@@ -18,6 +18,7 @@ import {
   focusedStorageFolderState,
 } from '../../recoilModel';
 import Home from '../../pages/home/Home';
+import { useProjectIdCache } from '../../utils/hooks';
 
 import { CreateOptions } from './CreateOptions';
 import { OpenProject } from './OpenProject';
@@ -38,12 +39,15 @@ const CreationFlow: React.FC<CreationFlowProps> = () => {
     updateCurrentPathForStorage,
     updateFolder,
     saveTemplateId,
+    fetchProjectById,
+    fetchRecentProjects,
   } = useRecoilValue(dispatcherState);
   const creationFlowStatus = useRecoilValue(creationFlowStatusState);
   const projectId = useRecoilValue(projectIdState);
   const templateProjects = useRecoilValue(templateProjectsState);
   const storages = useRecoilValue(storagesState);
   const focusedStorageFolder = useRecoilValue(focusedStorageFolderState);
+  const cachedProjectId = useProjectIdCache();
   const currentStorageIndex = useRef(0);
   const storage = storages[currentStorageIndex.current];
   const currentStorageId = storage ? storage.id : 'default';
@@ -57,9 +61,19 @@ const CreationFlow: React.FC<CreationFlowProps> = () => {
     }
   }, [storages]);
 
-  useEffect(() => {
-    fetchStorages();
+  const fetchResources = async () => {
+    // fetchProject use `gotoSnapshot` which will wipe out all state value.
+    // so here make those methods call in sequence.
+    if (!projectId && cachedProjectId) {
+      await fetchProjectById(cachedProjectId);
+    }
+    await fetchStorages();
     fetchTemplates();
+    fetchRecentProjects();
+  };
+
+  useEffect(() => {
+    fetchResources();
   }, []);
 
   const updateCurrentPath = async (newPath, storageId) => {
