+++ Composer/packages/client/src/pages/home/Home.tsx
@@ -3,7 +3,7 @@
 
 /** @jsx jsx */
 import { jsx } from '@emotion/core';
-import React, { useCallback, useEffect } from 'react';
+import React, { useCallback } from 'react';
 import formatMessage from 'format-message';
 import { Link } from 'office-ui-fabric-react/lib/Link';
 import { Icon } from 'office-ui-fabric-react/lib/Icon';
@@ -62,13 +62,9 @@ const Home: React.FC<RouteComponentProps> = () => {
   const recentProjects = useRecoilValue(recentProjectsState);
   const projectId = useRecoilValue(projectIdState);
   const templateId = useRecoilValue(templateIdState);
-  const {
-    openBotProject,
-    fetchRecentProjects,
-    setCreationFlowStatus,
-    onboardingAddCoachMarkRef,
-    saveTemplateId,
-  } = useRecoilValue(dispatcherState);
+  const { openBotProject, setCreationFlowStatus, onboardingAddCoachMarkRef, saveTemplateId } = useRecoilValue(
+    dispatcherState
+  );
 
   const onItemChosen = async (item) => {
     if (item && item.path) {
@@ -135,10 +131,6 @@ const Home: React.FC<RouteComponentProps> = () => {
       disabled: botName ? false : true,
     },
   ];
-
-  useEffect(() => {
-    fetchRecentProjects();
-  }, []);
 
   return (
     <div css={home.outline}>
